+++ lib/is_utf8.mli
@@ -1,0 +1,6 @@
+open! Core
+open! Import
+
+(** Heuristics for detecting files that are clearly not UTF-8 compliant. *)
+
+val clearly_not_utf8 : lines : string array -> bool